Remove required-projects from post docs jobs
The docs job is a final job, we cannot add required-projects here.

Remove them and the job, it's part of the publish template that is used.

Note that currently no post jobs are run due to the wrong setup. This
change breaks the post job temporarily but allows the other post jobs to
run while work is ongoing on the post docs job.

Change-Id: I373f870b47b69da7b034de2b7c4912bbd4a619aa