Remove legacy jobs from oaktree repos
The legacy jobs are actually broken currently, but the path forward for
them is to make nice new shiny ones. So go ahead and remove these.

Also, remove publish-to-pypi-quietly from oaktreemodel as oaktreemodel
will need a special build step before doing PyPI publication and that's
better suited for once we've done the base jobs.

Change-Id: I4c73e196b248d391549053beb7801bb8cb5e082a
Needed-By: Iee5f9b54d5c456ad5925f485b80f8e70accf962a