Fix publication of IRC meetings ICS files
During transition to Zuulv3, an error was introduced in the ICS
files publication post-job, resulting in the "calendars" directory
itself to be copied to $path/calendars, instead of the ICS files
present in the directory. ICS files were were therefore being
published to $path/calendars/calendars. This commit restores
publication to $path/calendars.

Change-Id: Ifebd883826341b4d4f21f2f114b0a87582845ab3