Move governance tag validation to experimental
The governance repository 'validate' test job currently only
tests that the diversity tags match the current demographics
in the project teams.

It is useful to validate diversity tag changes, but not so
much as a test job run on every change. Since diversity tags
are not continuously updated, it is expected that after a few
weeks the job will systematically fail.

It is still useful information in some cases, for example when
a new team is proposed and you're not sure which diversity tag
they should get, or when diversity tags are updated.

Move the validate test job to the experimental pipeline so that
the information can be retrieved in those specific cases by
running 'check experimental' as needed.

Change-Id: I420bd7bd568d102248dde43137a0a9273dfc6c41