Make stable jobs voting on Tempest
Stable jobs have always been voting on Tempest to prevent Tempest
changes from breaking stable branches. We never ran stable jobs on gate
though and I think that should stay the same. Given the low volume of
changes on stable, it's highgly unlikely that a change that passes in
check behaves differently on gate.

The voting part was lost in the zuulv3 migration, so restoring it.

Change-Id: Ifa69d4ea405f6c243eddf303cff7dcedec4bad4d