Publish deploy guide updates for all branches
The deploy guide for projects may change after the creation
of the stable branch, and those changes are not currently
publishing - instead leaving the deploy guide in the state
it was last published before the stable branch existed.

This changes the job to ensure that updates for all branches
are published, not just the master branch.

Change-Id: Id515a869ca43352a72953e7a2f50db364773b4ae