Don't index the devstack summary logs
The devstack summary logs are being picked up by logstash workers but
they are a proper subset of the full devstack log which we also index.
Avoid duplicative indexing by excluding the summary file.

This also addresses an issue with crm114 indexing these files with each
of them ending up with its own data dir due to timestamp data in the
filenames.

Change-Id: Ia397a3ceef0bf9f2198b4b6f9956d5b3dc9abcef