base/multinode rename #2: Rename project-config jobs
We want to rename the base and multinode integration jobs in order
to make sure they are not mistaken by users trying to test their
projects. These jobs are not meant to be used outside of integration
testing the playbooks and roles found in project-config, zuul-jobs and
openstack-zuul-jobs.

This is part two of three in renaming the base and multinode
integration jobs. We need to:
1) Add new jobs in openstack-zuul-jobs
2) Make project-config use the new job names  <-- We're here
3) Remove old jobs

Depends-On: I150a0a6dacc7b8862a40f1382ea730957dea0faf
Change-Id: I4ef44e64a03cc3089e02343de506e0a6fd85a55c