Expose serial keyword to base-test playbook
Now that zuul supports a single shared ansible_host in inventories,
add the ability for a user to use the serial keyword on our base
playbooks.

When using multi-host inventory files, all of the tasks happen
independent of each other on separate hosts. With a share ansibe_host,
they will end up fighting with each other. One way to avoid this, is
to set the serial keyword to 1, and host will run tasks one at a time.

By default, if base_serial is not set, we'll omit the serial keywork
in the playbooks.

Change-Id: I532939c843e7f6f6a7d29beb84ef523d25536bb3
Depends-On: I674d6baac26852ee1503feb1ed16c279bf773688
Signed-off-by: Paul Belanger <pabelanger@redhat.com>