Omit legacy npm job with nodejs4-publish-to-npm
When using the nodejs4-publish-to-npm project-template, its
release-openstack-javascript job is redundant with the
legacy-npm-upload job so remove the latter. Also in the case of
monasca-grafana-datasource there's no need for
publish-openstack-python-tarball since it doesn't work on that
project and was only added to satisfy (potentially invalid)
dependencies of the legacy-npm-upload job.

Change-Id: I7f1e98d0caee791271dcb08bbf4a76154fb065a3