releasenotes: Remove package install
The vast majority of changes to enable running releasenotes without
installating of the package itself are merged. Remove the package
install now. Note that install-if-python is the proper role for this,
the previous setting was a noop.

The remaining few changes can merge at any time, projects had enough
time.

See also
http://lists.openstack.org/pipermail/openstack-dev/2017-November/124815.html

Change-Id: I225d73a2f9321329b34a282e08aea6773622df5b
Depends-On: I6d461a69537d3f7dd44e9b9877f652df656eedd8