Fix npm publishing
During the conversion, the nodejs4-publish-to-npm template was replaced
with python publishing jobs by accident.

Really use nodejs4-publish-to-npm for the publishing and remove thus
legacy-npm-upload and publish-openstack-python-tarball for npm repos.

Remove also npm-tarball-upload.sh, it is only used by the
legacy-npm-upload job and thus not needed anymore.

For reference, see
http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml?h=before-jenkins-config-removed
and check the templates used by the projects changed.

Needed-By: I696a23792f8e4056b788ad4a8b018c423aa1af68
Change-Id: I5cccd00518afcc50c575a4bb1937345a30fe584f