Create the "ara-web" project
One of the features in the upcoming release of ARA is the separation
and clear physical and logical boundaries between the different
components such as:
- Core (backend/model/API)
- Web (web application interface)
- Callbacks (Ansible callbacks and their dependencies)
- Clients (CLI and API clients)

This allows users to install what they need without being burdened by
the dependencies of components they are not interested in.
For example, one might want to use the callbacks somewhere without
having to install all the web application dependencies.

Users installing "ara" will still pull everything but the opportunity
to install only what is necessary will be available.

The "ara-web" project is the first of those components being split out.
There are no upstream to import, we will start from a clean repository.

Change-Id: Ic46bff4a152560d787d40bb10ee94a6abee16485