Add job for publishing built javascript content
There are two different types of tarballs we want to publish, a source
tarball to npm, and a built tarball intended for deployment. We have a
job for dealing with tarballs to npm, but we're lacking a new-style job
for publishing built artifacts to tarballs.o.o.

Add one, and switch storyboard to using it.

Change-Id: I19607d14facb93d1f46ea68feba563badfbbb5a2