Small refactor of translation scripts
The modulename parameter for setup_projects are not needed anymore,
their usage has been removed with retirement of mitaka. Remove these
parameter.

Call setup_project only once per project and not for both python and
django setup. Also, move the zanata pull out to not have to create a
lock.

Change-Id: Iaf58a6882527eee3ce6528b516fe0f24f51cd976