Drop django_openstack_auth master branch jobs
In Queens cycle, the code of django_openstack_auth was merged into
the horizon repository. The master branch of django_openstack_auth
will be retired.

This commit skips the jobs of django_openstack_auth for the master
branch. We still maintain the stable branches, so the jobs need to be kept.
Zuul project template does not support brances, so required project
templates are converted into jobs.

system-required template is kept as this is just a merge-check noop job.
publish-to-pypi is kept too as we don't release django_openstack_auth
from the master branch and there is no side effect.

Change-Id: I837c3426de1a18b4055dd7066743e427074425b3