Add neutron-lib-commit group
We need to increase the number of people with W+ power in the
neutron-lib project to keep up with the number of patches being
submitted. neutron-lib-core, that has +2 power, includes all the neutron
stadium projects and would be too big a group. This patch adds a
neutron-lib-commit group to have finer control on who has W+ power

Change-Id: Ia5daabe1a48a4630577e678dbfd8d1593fafcefb