Fix grenade docs jobs
Grenade needs to run "tox -e docs", just using sphinx is not enough.
Use the tox-docs job for building since grenade does not use
requirements.

Reuse the i18n publish job and rename it so that it can be used for both
repositories - it does already what we need.

Note that the current publish job is broken, last update to
https://docs.openstack.org/grenade/latest/ was in July.

Needed-By: Ibbc60d3d6e66b7f6a850ebd9268c5b5b36c37baa
Change-Id: I6743079d4908fb4585c3714efcd0af7d4b404588