add neutron-lib master periodic job to vmware-nsx
We currently run the periodic-jobs-with-neutron-lib-master on a number
of sub-projects that consume neutron-lib. These periodics act as a
smokescreen to determine consumer health and is used to determine if
it's safe to release neutron-lib.

This patch adds vmware-nsx to that set of consumers by including the
periodic-jobs-with-neutron-lib-master template on the vmware-nsx
project. It also adds the periodic results to the neutron-lib
grafana dashboard.

Change-Id: Ic94010417fc702e2ebf62e29bb29c363af68d10c