Publish neutron and horizon to pypi
The suggestion has been made to publish the server projects to PyPI:

  http://lists.openstack.org/pipermail/openstack-dev/2017-November/124676.html

Doing this will help the neutron and horizon plugin projects not need to
do special things WRT dependencies.

There is an issue to be solved before we can do it for everything.
Keystone is currently https://github.com/dcrosta/keystone on PyPI, not
us. An email has been sent to the maintainer of that package, which
seems very unactive, but no response has happened.

For now, just start publishing neutron and horizon. We already have an
old horizon up there anyway, and we have the neutron entry. If we
approve this, we can also write a one-off script to upload the
older/existing releases from tarballs.o.o.

Change-Id: Ia8cb00045aee3ce8cfc506309ab35477a2d686a1