Add SQL reporter to Zuul experiemental and tripleo pipelines
The Zuul MySQL database is one of the only ways we can get fairly
precise historical metrics for the jobs that run in all the pipelines.

However, we're missing a significant amount of data right now by
leaving out the experiemental, check-tripleo and
check-tripleo-experimental pipelines.

Let's start recording those so we can make informed decisions with
numbers easily in the future.

Change-Id: I0daeff1f6c6d4527cbd17f55c2df242e1ecb8b33