Disable infracloud-chocolate on nl02.o.o
Our controller in infracloud-chocolate has stopped replying to
requests. Currently we don't know what is happening, and access to iLo
card for the server isn't responding.

It maybe possible the hardware was pawered off or failure of HDDs. For
now, disable the provider and see if we can get any feedback from HPE.

Change-Id: Iffb7bed95b5e4ab9be17cb210b5f78e445cdab1c
Signed-off-by: Paul Belanger <pabelanger@redhat.com>