Temporarily reduce max-servers in rax-dfw
We're seeing an odd rate limiting behavior on our mirror server in
rax-dfw. Ticket 180130-ord-0000697 has been filed with the provider
to investigate, but while that is underway we'll reduce our
max-servers in that region from 140 to 100 to reduce network
bandwidth consumption on it and relieve the resulting packet loss
and package retrieval slowness/timeouts/errors.

Change-Id: I1ab72262bd2f9f725188a2546f92c7f515c762cc