Fix vmware-nsx stable-periodic docs job
The vmware-nsx job has been failing with an error looking for
the networking-l2gw repo. Turns out there are several repos
needed by this job, but the stable-periodic pipeline did not
define the required-projects that were used elsewhere.

This adds the required-projects used in the check and check
and gate queues to the job in the periodic-stable pipeline.

Change-Id: I29a7dd501ec9878b7b9bd14f1bd2db957e78e142