Switch static-special from copy to synchronize
According to notes in the Ansible documentation for the copy module,
it "does not scale to lots (>hundreds) of files." Several of the 5
sites currently relying on the publish/static-special playbook have
content numbering in the thousands of files. Other sites on the same
server instead rely on the publish/static playbook which already
uses synchronize instead of copy, so this should be entirely
compatible.

Change-Id: Iac85d906ae7c7a4666d7a5eeaa73b5a9f2ac7115