Switch horizon projects to publish-to-pypi-horizon
While checking some stable release changes, it was found there
are a few horizon plugin projects that have not been updated
since the changes with zuulv3. There are a few projects that
either require horizon to be installed, or have not yet
explicitly declared support for the normal publish-to-pypi
release job.

Until they state otherwise, we need to switch these projects
over to the -horizon flavor of this job to be able to pass
validation.

Change-Id: I8bc5bc8101d8375ab20b7e07dd1bfb2f087b20d7