Remove legacy-devstack-dsvm-py35-updown from devstack
We are currently running both python3.5 and 2.7 versions of the
"updown" test (this brings up devstack, then runs ./unstack.sh as a
closing hook).

We run legacy-tempest-dsvm-py35 as part of integrated-gate-py35; so
this python3.5 updown job isn't really exercising anything different,
and mostly just duplicates the 2.7 updown job.

As we're moving everything to native jobs slowly anyway, I think we
can just remove the python3.5 job to simplify that.

Change-Id: I62b2ca20a3ed0f1938c1212a5d2652e513a7cfce