Add vault charm to gerrit
Add the vault charm to the charm project and to gerrit.
This includes a new acl-config file for the vault charm, this is to
allow the existing devs working on the vault charm on LP to continue
to be able to review and land changes, without giving them access to
approve changes to all OpenStack Charms.

Change-Id: If5701b821060c1b666c102d791735fb14cb15e68