Disable citycloud in nodepool
At the request of Daniel Öhberg, CityNetwork VP of Operations, we
are ceasing booting CI nodes in any remaining regions and pausing
all image uploads for now. We should get back in touch with them
around the beginning of July to find out whether they're ready for
us to reenable use of this account.

Change-Id: I39f12d5a8441b197f5a310c7b33ba42c04c73378