Don't run legacy-tempest-dsvm-neutron-full for nova on queens+
The integrated-gate job template in openstack-zuul-jobs has the
proper branch restriction on legacy-tempest-dsvm-neutron-full
so that it won't run on stable/queens+ changes. However, the
override of the job in project-config for nova means we end up
running that job and its replacement (tempest-full) on queens+
changes.

So this change removes the override so nova changes get the job
definition, including branch restrictions and irrelevant-files,
from the template in openstack-zuul-jobs. This does mean we could
be running this job on stable/ocata or stable/pike nova changes
that don't really need it, but that should be rare.

Change-Id: Ib5d12dbdc69a0506ea53fd130ccf816557503bae