Move jenkins/data/bindep-fallback.txt (1/2)
Let's get rid of jenkins in the name and use a new place to place
bindep-fallback.txt in.

Create new nodepool element to copy the file over.

This is a first of two changes for this repo, we also need to update
the few users in bindep and openstack-zuul-jobs before we can remove the
old location.

We can remove jenkins/data/bindep-fallback.txt only in the followup
change since project-config is not self-testing.

Change-Id: I5c94dcbbca25db7dd998bfbccfbb25c1c4342a2b