Move jenkins/data/bindep-fallback.txt (2/2)
Remove using old location, we can switch now to the new location
everywhere.

Note: We can only merge this after new images are build with the new
location build in.

This needs a recheck once change 1 is in since project-config is
not self-testing. So, we can remove jenkins/data/bindep-fallback.txt in
this change.

Change-Id: I08f23e96defb152946ade1490628f37f0215300d
Depends-On: https://review.openstack.org/543141
Depends-On: https://review.openstack.org/543142