Rework base-publish-static jobs using protected
The new protected attribute means that only jobs in the same repo can
inherit from it. Use that on base-publish-static and move the secrets
from the inherited job to it now, it's now safe to have them here.

Since the secrets need to be available to the post-job and are not
available to post-jobs of inheried jobs, use a common post job for all
of these. Move static-special.yaml to static.yaml and set the location
for the publish location for the single previous user of static.yaml.

Change-Id: I3d749adf242eb87895ad10b6f9c4d12859882ac9