Refactor playbooks/publish/openstack-afs.yaml
Rework the playbook so that it takes an additional parameter and we can
use it in more places (see followup change). Set the current value
as default

Note that we really need to allow a value of "{{ afs.path }}" for the
variable without trailing "/" so that publishing of openstack-manuals
works which publishes to the root folder.

Change-Id: I15b77fc002a6b099ff23e0c07972b49ad743a2c9