Fix ceph failure rate graph
The gate-tempest-dsvm-neutron-full-ubuntu-xenial
job was renamed to tempest-full in Queens during
the zuulv3 transition.

The ceph job was renamed to
legacy-tempest-dsvm-full-devstack-plugin-ceph
during the zuulv3 transition.

Obviously this graph is going to be wrong for any
of the old job names, or results before the job
renames, but since the ceph job is non-voting
it's likely only really useful to see the results
since queens anyway.

Change-Id: Ie00a637cfcd718320631da3357794f7d97f31b1e