Import config_template to openstack(-ansible)
This patch adds the config_template action plugin, coming from
OpenStack-Ansible plugins, but spined off as standalone to get
more community feedback. It's a truely generic plugin that
can be used everywhere in ansible.

An individual core group is added, but the general OSA core
group would have core capabilities as a back-stop for now, until
more diversity is found.

Needed-By: I7cbf22fff1c483f23c59cae4d9e6669f4c857f04
Change-Id: Icf63d99016c8c623a9e82b2832520c5751b3a2cd