Regex gerritbot notifications: #openstack-nova
Since [1] we support regular expressions syntax in channel configuration
YAML files for the gerritbot.

This patch applies regexes for the channel listed in the title to:
- reduce the size and improve readability of the paragraph
- eliminate the need for subsequent patches as new stable/* branches are
  published.

Note that we're not bothering to exclude old/EOL stable branches.  The
philosophy here is that notifications for those branches should almost
never happen anyway; and if they do, you want to know about them.

[1] https://review.openstack.org/499367

Change-Id: Ice6fa52cbfaf3fa7d3a1dd209c24b718efc0ba29