Switch notifications for stable branches to match all
With the ability to use regex matching, we can now set one
default stable pattern to match. Since the presence of old
deleted branches are still in the config, it would appear
projects are not always keeping the individual names up to
date. This switches them all to just match on any stable
branches to get rid of the need to update these each
release.

Change-Id: I747b65a9422d8e254ab690b5390b7bf5cd517f6b