Switch astara release jobs to not use pypi
There were a few astara repos that were setup up with the
publish-to-pypi release job. None of the astara projects
has anything registered on pypi though, so release job
verification fails for these on trying to validate that
openstackci has rights to push up a new release.

This project is basically retired, so the only reason we
need to change these now is to be able to make other
updates to the legacy release deliverable files in the
releases repo, since changes to those files cause validation
to run.

Change-Id: Id2482015c6920670e3cfe19581f35df99f41ab07