Make combined and correct acls for openstacksdk team
The shade and sdk teams have merged, but the original implementation of
this in the ACLs was to just make openstacksdk-core include shade-core.
os-client-config has also been merged, and requestsexceptions is a
deliverable of the team too.

Go ahead and collapse them into a single acl file and apply it to all
the projects. Also remove the -release team as openstacksdk is release
managed.

Update the ACL file to match what it should be for a release-managed
project, including stable-maint and branch-release teams.

Update the description - that wiki page isn't really a thing we
care about.

Change-Id: I45364a382fbed7bbadbecd29aa6ff051fecdb895