Add ubuntu-bionic to nodepool
As ubuntu-bionic is the new LTS for ubuntu, start providing images in
nodepool.  Something to note, bionic isn't actually released yet.
However our tooling does work and is able to create workable images.

We should hold off to migrating jobs from ubuntu-xenial to
ubuntu-bionic until it is officially released, and we have a migration
plan in place.

Change-Id: If12435969ca37400bfd04abe8470048146be7aa3
Signed-off-by: Paul Belanger <pabelanger@redhat.com>