Stop requests-mock jobs for retirement
requests-mock was always a general purpose python library, however as it
was initially developed in the openstack community it was hosted there
to take advantage of the infrastructure.

At this time with its developers spending less time on OpenStack it is
felt that being behind gerrit is limiting contributions (even without
the CLA) as more outside contributions are received by launchpad
attachment than gerrit.

The project will continue development at
https://github.com/jamielennox/requests-mock

Remove the zuul jobs in accordance with:
https://docs.openstack.org/infra/manual/drivers.html#retiring-a-project

Change-Id: I7e88b9fc68f83b8a37b99e62f207dbd648a114e4