Remove legacy-tempest-dsvm-cells job usage
The cells v1 job definition is being moved to nova and renamed
to 'nova-cells-v1'. The job usage is now defined in nova, devstack
and tempest repos so we can remove the legacy job name usage from
project-config.

Note that the related per-repo patches will need to be backported
to stable/queens (except tempest) to run the job there (the cells v1
job is currently blacklisted from stable/pike and ocata, so we're not
worrying about those branches right now).

Depends-On: https://review.openstack.org/549780
Depends-On: https://review.openstack.org/549781
Depends-On: https://review.openstack.org/549783
Change-Id: I5cfd9890856d289b23f79a3d1c1c01f0376ec897