Create new project for oslo.limit
During the Rocky PTG, it was decided that services should have some
sort of library to help with quota enforcement based on limits stored
in keystone. This relationship is very similar to how services
enforce policy decisions based on role assignments in keystone using
the oslo.policy library.

This commit creates a new oslo library called oslo.limit to fill this
need.

Needed-By: I46d5d16f29a4a613c96f01023c89e94b39abf2dd
Change-Id: I1bdb05189fa0346426cb36d4a3d5fe458eb69871