gerritbot: send kolla/kolla-ansible stable notifications to #osism channel
OSISM is a downstream of kolla/kolla-ansible. To improve our workflow,
notifications to the stable branches should be sent to our IRC channel.
We can react better by making the changes more visible to us.

Due to the high activity in #openstack-kolla and notifications to master
on kolla/kolla-ansible/kolla-kubernetes it is very impractical for
us there.

Change-Id: I037fd07abb32659ad494f67ee9b43c25149db799