Bring vexxhost back online for testing
We created a new mirror in vexxhost with no bandwidth limits, the
thought here is this should results in better job run times as our
mirror will not be the bottleneck.

We've also changed flavors to use v2-highcpu-8 which are 8C/8Gb
memory, but require us to now boot-from-volume with 80GB volume-size.

Change-Id: Ibb26597cfd3c3b10891d7a60e70c6143d1c12407
Signed-off-by: Paul Belanger <pabelanger@redhat.com>