Create the ara-server project
This follows up on the change to create the ara-web project [1] and the
split of the ARA components in the upcoming release.

We are splitting the components in order to allow indenpendant
installation and deployment for each component.

The ara-server project will contain the backend/model and API for ARA.

[1]: https://review.openstack.org/#/c/528358/

Change-Id: I9d40e00588b03da6aca55e1ad1c7009635f3e4ec