Create the ara-clients project
This follows up on the change to create the ara-web project [1] and the
split of the ARA components in the upcoming release.

We are splitting the components in order to allow indenpendant
installation and deployment for each component.

The ara-clients project will contain the client interfaces, modules
and libraries for ARA.

[1]: https://review.openstack.org/#/c/528358/

Change-Id: Ia7a413e596bc4291cde1e342caf1d1c2c7b719fe