Create the ara-plugins project
This follows up on the change to create the ara-web project [1] and the
split of the ARA components in the upcoming release.

We are splitting the components in order to allow indenpendant
installation and deployment for each component.

The ara-plugins project will contain the plugins, modules and callbacks
for ARA.

[1]: https://review.openstack.org/#/c/528358/

Change-Id: I2d345414d16d4f08f9d4570a92f24b7a43203e66