Bump max-servers limestone nodepool provider to 10
We've manually and successfully tested a virtual machine with the
nodepool image and the ipv6 network.
The virtual machine could reach the regional mirror and a devstack-gate
reproduce.sh ran to completion.

Let's do an initial bump to 10 nodes to monitor for failures we could
not catch -- we'll bump to 50 afterwards if we're confident things are
working well for both us and Limestone.

Change-Id: I3a3b68dd9fbd0243f3544ceceb4859aa6d435fc8