Catch errors in release processing script
The releases playbook calls process_release_requests.sh, which
calls process_release_requests.py, which calls release.sh and
make_branch.sh as necessary.

If an error happens during release.sh or make_branch.sh, the error
is propagated back to process_release_requests.py, but that script
failed to propagate the error back to process_release_requests.sh.

This could result in the job appearing to complete successfully,
while it actually failed to do anything (whatever the reason is).

Change-Id: I10e5362e4d14cb9d787cf09f62d0b67f7dce8a03