check out the series branch when tagging a release
When a project is retired, the .gitreview file is removed. Without
that file, "git review -s" (run during cloning) fails, and the tag job
ends up failing.

This patch passes an assumed branch name to clone_repo.sh so that if
the stable branch for the series exists we get it. If the stable
branch does not exist, clone_repo.sh falls back to checking out
master.

Change-Id: Ib1221b5672c518eb8e3037232440807af6c90412
Signed-off-by: Doug Hellmann <doug@doughellmann.com>