base-test: Gather network facts for all hosts in pre-run
Today, we do this in the validate-host role, however because we need
to access the ansible network variables in configure-unbound, that is
too late in our task process.

Because we have a fact cache by default, we can just enable the
gathering of facts when we run our first pre-run playbook, then our
validate-host role to gather then becomes a noop.

Change-Id: I9b56b3cbfd45f45a8e0d4477737f94fb0babc97c
Signed-off-by: Paul Belanger <pabelanger@redhat.com>