Revert "Stop systemd-resolve from booting on our images"
This reverts commit da1552359561a35b802dde534073f2ea5fec6372.

systemd-resolved was not really involed in this.  What seems to be
happening is that during the initial chroot creation there is no
/etc/resolv.conf file, and so the systemd postinst script creates one
pointing to the compatability files created by systemd-resolved.

This is not what we want, but dib doesn't really provide a way for us
to overwrite the file.  That is covered in
Ie0e97d8072e2b21a54b053fa6fb07b62960c686d

We actually want systemd-resolved running -- it provides the dbus
nameserver api stuff that some tools may use.  If /etc/resolv.conf is
a file, systemd-resolved leaves it alone and uses it for resolution as
you would expect.  (we do, however, want to wait for the depends-on to
ensure the image has a correct /etc/resolv.conf before merging this).

Depends-On: https://review.openstack.org/557842/

Change-Id: Ie3cdc323bf8fb4fcee725f9a52bf53a4a6a01bbf