Remove ansible-role-functional-jobs-fedora-26 for windmill
Because we are removing fedora-26 support from nodepool, remove these
jobs so we don't block that effort. This is fine, because I haven't
found the time to actually update these roles to test against
fedora-27.

And because I want to move these jobs in-tree, this give me the
modivation to add them back as fedora-27.

Change-Id: I5cf5a2d4e14b54f69bb250bf4a0b3053d54b3fa1
Signed-off-by: Paul Belanger <pabelanger@redhat.com>