Add entries for #openstack-placement IRC channel
An IRC channel for openstack-placement has been created to cut down
on the amount of placement-only traffic in #openstack-nova. This
adds gerritbot and accessbot entries for the new channel, which
has been registed and granted for 'openstackinfra' as described in
https://docs.openstack.org/infra/system-config/irc.html

A subsequent patch in system-config will add support for meetbot
and statusbot.

Change-Id: Ib5e9f7252343b1ced0ad482c516ce5c17652b64e