Remove direct tag / branch ACLs for official projects
The TC confirmed that creating tags (or branches) for repositories
in official deliverables should be handled by the release management
team through the openstack/releases repository.

Yet some ACLs still allow for direct tagging or creating references
by some other teams, which can cause process disruption.

This removes tagging and branch-creating rights for official
deliverables that are already handled by the release management team,
or which never released in the past.

The tool that generated this patch (and contains the list of remaining
team and repository exceptions to the rule) will be posted to the
releases repository.

Depends-On: https://review.openstack.org/557737
Change-Id: I9feea4a59fccec940a36a7953d35a6f734d86582