Updates jobs running on the election repo
Currently we run the requirements-check job on the election repo which
isn't really helpful as these tools don't really need to be
co-installable with OpenStack services.

We're also *not* running any kind of linter so we have a few code
style issues creeping in.

This removes requirements-check and adds a non-voting linter to the
check pipeline.  We can upgrade the latter to voting once all issues
are fixed, which wont be during the current TC election.

Change-Id: Id2ad76b587ef50104d6410c8d6c497bf291ed662