Stop using -python35-jobs-horizon for trove-dashboard
We trie to make the py35 jobs temporarily disabled for stable branches
of trove-dashboard in I3421c945534dc249c42cef9977398ce6c376436e,
however it was not working as expected. Recently I found monasca-ui did
a similar change and worked, the difference is that monasca-ui is not
using the openstack-python35-jobs-horizon template, so the problem
should be the 'branches' attribute defined in projects.yaml cannot be
applied the jobs inheritted from project templates.

Change-Id: I34984c0c1942686d751e59adf37b920b66bd15cd
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>