create constellations repository
The TC is starting a new documentation initiative to help deployers
understand how to combine various OpenStack components for different
purposes. Although the TC is starting this, ultimately we hope to
build a team of maintainers for the repository and so a new gerrit
group is being used as the reviewers. We do not anticipate tagging or
branching, so those permissions are left as the defaults. We don't
know yet if this will remain a "TC-owned" repo or if the documentation
team will adopt it, so there is not yet a related governance patch.

Story: #2001702
Task: #15069
Change-Id: I0f7dccd1503b59d0e61eec00913bd96bf481184d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>