Remove legacy-tempest-dsvm-multinode-live-migration job usage
The live migration job definition is being moved to nova and renamed to
'nova-live-migration'. The job usage is now defined in nova
and tempest repos so we can remove the legacy job name usage from
project-config.

Note that the related per-repo patches will need to be backported to
stable branches to run the job there.

Depends-On: https://review.openstack.org/555945
Depends-On: https://review.openstack.org/565608

Change-Id: I8f7b62d96320ce71a4bfa25c2d9a36e973bf6eb6