New Airship project - Diving Bell
Airship is a collection of components that provide a declarative
interface for deploying a bare metal Kubernetes cluster at scale to
facilitate integrated deployment of OpenStack on Kubernetes using the
existing OpenStack-Helm project [1].

Diving Bell is a lightweight host configuration management tool.  It
fits in the Airship ecosystem in areas where leveraging containers for
configuration management is not ideal.

[1] https://github.com/openstack/openstack-helm

Change-Id: I7728db62cfa2400808307958fc9e7e061ae2a04d