Add initial Airship projects
Airship is a collection of components that provide a declarative
interface for repeatable deployment and operation of OpenStack on top of
Kubernetes using the existing OpenStack-Helm project [1].

The following projects are added with this change:

Armada - An orchestrator for deploying and upgrading a collection of
Helm charts.

Berth - A lightweight mechanism for managing VMs on top of Kubernetes
via Helm.

Deckhand - A configuration management service with features to support
managing large cluster configurations.

Drydock - A declarative host provisioning system built initially to
leverage MaaS for baremetal host deployment.

MaaS - A collection of artifacts needed to deploy MaaS on top of
Kubernetes.

Peg Leg - A tool to organize configuration of multiple Airship
deployments.

Promenade - A deployment system for resilient, self-hosted Kubernetes.

Shipyard - A cluster lifecycle orchestrator for Airship.

Airship-in-a-bottle - Integrated configuration and documentation.

[1] OpenStack Helm: https://github.com/openstack/openstack-helm

Change-Id: I68e58ec68eacdd9a9d3f25b07082429dbfef668d